refactor(client): organize pages into folders
