+++ client/src/ArtistPage/Artist.styled.js
