+++ client/src/ArtistPage/Profile.jsx
