+++ client/src/ArtistPage/Profile.styled.js
