+++ client/src/Artists/Artists.jsx
@@ -1,19 +1,19 @@
 import React from 'react';
-import HomeWrapper from './Home.styled';
+import ArtistsWrapper from './Artists.styled';
 import Card from './Card';
-import Data from './data';
+import Data from '../data';
 
 // @todo make this a class, fetch data on mount
 const Bands = Data.bands;
 
-const Home = () => (
-  <HomeWrapper>
+const Artists = () => (
+  <ArtistsWrapper>
     {Bands.map(band => (
       <Card key={band.name} bandUrl={band.url} image={band.heroImage}>
         {band.name}
       </Card>
     ))}
-  </HomeWrapper>
+  </ArtistsWrapper>
 );
 
-export default Home;
+export default Artists;
